+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add `kp gh-release` command to publish a new release on GitHub.
+- Migrate this package to esm format.
 
 ## 0.9.7